refactor: order entity 생성하여 로직 분리
