+++ src/basic/components/SummaryDetails.js
@@ -1,0 +1,72 @@
+export const SummaryDetails = ({
+  totalOriginalPrice,
+  cartItems,
+  prodList,
+  totalItemCount,
+  itemDiscounts,
+  isTuesday,
+  totalDiscountedPrice,
+}) => {
+  let summaryDetails = document.getElementById("summary-details");
+  summaryDetails.innerHTML = "";
+  if (totalOriginalPrice > 0) {
+    for (let i = 0; i < cartItems.length; i++) {
+      let curItem = null;
+      for (let j = 0; j < prodList.length; j++) {
+        if (prodList[j].id === cartItems[i].id) {
+          curItem = prodList[j];
+          break;
+        }
+      }
+      const qtyElem = cartItems[i].querySelector(".quantity-number");
+      const q = parseInt(qtyElem.textContent);
+      const itemTotal = curItem.price * q;
+      summaryDetails.innerHTML += `
+        <div class="flex justify-between text-xs tracking-wide text-gray-400">
+          <span>${curItem.name} x ${q}</span>
+          <span>₩${itemTotal.toLocaleString()}</span>
+        </div>
+      `;
+    }
+    summaryDetails.innerHTML += `
+      <div class="border-t border-white/10 my-3"></div>
+      <div class="flex justify-between text-sm tracking-wide">
+        <span>Subtotal</span>
+        <span>₩${totalOriginalPrice.toLocaleString()}</span>
+      </div>
+    `;
+    if (totalItemCount >= 30) {
+      summaryDetails.innerHTML += `
+        <div class="flex justify-between text-sm tracking-wide text-green-400">
+          <span class="text-xs">🎉 대량구매 할인 (30개 이상)</span>
+          <span class="text-xs">-25%</span>
+        </div>
+      `;
+    } else if (itemDiscounts.length > 0) {
+      itemDiscounts.forEach((item) => {
+        summaryDetails.innerHTML += `
+          <div class="flex justify-between text-sm tracking-wide text-green-400">
+            <span class="text-xs">${item.name} (10개↑)</span>
+            <span class="text-xs">-${item.discount}%</span>
+          </div>
+        `;
+      });
+    }
+    if (isTuesday) {
+      if (totalDiscountedPrice > 0) {
+        summaryDetails.innerHTML += `
+          <div class="flex justify-between text-sm tracking-wide text-purple-400">
+            <span class="text-xs">🌟 화요일 추가 할인</span>
+            <span class="text-xs">-10%</span>
+          </div>
+        `;
+      }
+    }
+    summaryDetails.innerHTML += `
+      <div class="flex justify-between text-sm tracking-wide text-gray-400">
+        <span>Shipping</span>
+        <span>Free</span>
+      </div>
+    `;
+  }
+};
